Final Commit
I have finished writing the test that I believe would have solved the
second best buy user story.

In order to finish this story, I would probably need to create a new
class method in the Store model. This new class method would use another
method in the best buy service that would search for that particular
store's information.

This might get a little tricky because I wasn't sure if the long name,
which is currently coming through in params, would be sufficiently
unique to run the search.

If not, I would probably need to go back to the "find all by zip" method
and have it grab the id of the store, which would be sufficient. In
fact, now that I'm writing this, I realize that's what I should probably
go back and do anyway.